+++ TabletopTweaks/NewContent/Feats/ShatterDefenses.cs
@@ -10,10 +10,10 @@ using TabletopTweaks.Utilities;
 
 namespace TabletopTweaks.NewContent.Feats {
     static class ShatterDefenses {
-        public static void AddShatterDefensesBuffs() {
+        public static void AddNewShatterDefenseBlueprints() {
             var ShatterDefenses = Resources.GetBlueprint<BlueprintFeature>("61a17ccbbb3d79445b0926347ec07577");
 
-            var ShatterDefensesDisplayBuff = Helpers.CreateBuff("ShatterDefensesBuff", bp => {
+            var ShatterDefensesDisplayBuff = Helpers.CreateBuff("ShatterDefensesDisplayBuff", bp => {
                 bp.m_Icon = ShatterDefenses.m_Icon;
                 bp.Stacking = StackingType.Prolong;
                 bp.SetName("Shattered Defenses");
@@ -42,23 +42,6 @@ namespace TabletopTweaks.NewContent.Feats {
                 bp.SetName("Shattered Defenses");
                 bp.SetDescription("An opponent you affect with Shatter Defenses is flat-footed to your attacks.");
             });
-            var ShatterDefensesMythicBuff = Helpers.CreateBuff("ShatterDefensesMythicBuff", bp => {
-                bp.m_Icon = ShatterDefenses.m_Icon;
-                bp.SetName("Shattered Defenses (Mythic)");
-                bp.SetDescription("An opponent affected by Shatter Defenses is flat-footed to all attacks.");
-            });
-            var ShatterDefensesMythicFeat = Helpers.CreateBlueprint<BlueprintFeature>("ShatterDefensesMythicFeat", bp => {
-                bp.m_Icon = ShatterDefenses.m_Icon;
-                bp.SetName("Shatter Defenses (Mythic)");
-                bp.SetDescription("An opponent you affect with Shatter Defenses is flat-footed to all attacks, not just yours.");
-                bp.IsClassFeature = true;
-                bp.Ranks = 1;
-                bp.Groups = new FeatureGroup[] { FeatureGroup.MythicFeat };
-                bp.AddPrerequisiteFeature(ShatterDefenses);
-            });
-            if (ModSettings.Fixes.Feats.IsDisabled("ShatterDefenses")) { return; }
-            if (ModSettings.AddedContent.MythicFeats.IsDisabled("MythicShatterDefenses")) { return; }
-            FeatTools.AddAsMythicFeat(ShatterDefensesMythicFeat);
         }
     }
 }